refactor(clinics): move use cases into clinics folder
