+++ src/main/java/com/medspace/application/usecase/clinic/photo/AssignPhotoToClinicUseCase.java
@@ -1,4 +1,4 @@
-package com.medspace.application.usecase.clinicPhoto;
+package com.medspace.application.usecase.clinic.photo;
 
 import com.medspace.application.service.ClinicService;
 import com.medspace.domain.model.ClinicPhoto;
@@ -8,19 +8,17 @@ import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
 
 @ApplicationScoped
-public class DeleteClinicPhotoByIdUseCase {
+public class AssignPhotoToClinicUseCase {
     @Inject
     ClinicService clinicService;
 
     @Transactional
-    public void execute(Long photoId, Long userId) {
-        ClinicPhoto clinicPhoto = clinicService.getClinicPhotoById(photoId);
-        Long clinicId = clinicPhoto != null ? clinicPhoto.getClinic().getId() : null;
+    public ClinicPhoto execute(Long clinicPhotoId, Long clinicId, Long userId) {
         Boolean isOwner = clinicService.validateClinicOwnership(clinicId, userId);
         if (!isOwner) {
-            throw new ForbiddenException("Delete unauthorized");
+            throw new ForbiddenException("Assign unauthorized");
         }
 
-        clinicService.deleteClinicPhotoById(photoId);
+        return clinicService.assignPhotoToClinic(clinicPhotoId, clinicId);
     }
 }
